Rename schemas/hunts -> hunt
