+++ imports/lib/models/hunts.ts
@@ -1,8 +1,8 @@
-import HuntsSchema, { HuntType } from '../schemas/hunts';
+import HuntSchema, { HuntType } from '../schemas/hunt';
 import Base from './base';
 
 const Hunts = new Base<HuntType>('hunts');
-Hunts.attachSchema(HuntsSchema);
+Hunts.attachSchema(HuntSchema);
 
 // All hunts are accessible, since they only contain metadata
 Hunts.publish();
