+++ imports/model-helpers.ts
@@ -3,7 +3,7 @@ import { Mongo } from 'meteor/mongo';
 import { _ } from 'meteor/underscore';
 import Mustache from 'mustache';
 import MeteorUsers from './lib/models/meteor_users';
-import { HuntType } from './lib/schemas/hunts';
+import { HuntType } from './lib/schemas/hunt';
 import { PuzzleType } from './lib/schemas/puzzles';
 
 const answerify = function (answer: string): string {
