+++ imports/server/guesses.ts
@@ -9,7 +9,7 @@ import Profiles from '../lib/models/profiles';
 import Puzzles from '../lib/models/puzzles';
 import { userMayUpdateGuessesForHunt, deprecatedIsActiveOperator } from '../lib/permission_stubs';
 import { GuessType } from '../lib/schemas/guess';
-import { HuntType } from '../lib/schemas/hunts';
+import { HuntType } from '../lib/schemas/hunt';
 import { PuzzleType } from '../lib/schemas/puzzles';
 import { sendChatMessage } from './chat';
 import GlobalHooks from './global-hooks';
